Move ProjectModel, ServiceModel to :apollo-tooling
