+++ libraries/apollo-compiler/api/apollo-compiler.api
@@ -1948,73 +1948,6 @@ public final class com/apollographql/apollo/compiler/ir/IrVariable$Companion {
 	public final fun serializer ()Lkotlinx/serialization/KSerializer;
 }
 
-public final class com/apollographql/apollo/compiler/model/ProjectModel {
-	public static final field Companion Lcom/apollographql/apollo/compiler/model/ProjectModel$Companion;
-	public fun <init> (Ljava/util/Set;Ljava/lang/String;Ljava/lang/Integer;Ljava/lang/Integer;Ljava/lang/String;Ljava/lang/String;Ljava/lang/Boolean;Ljava/lang/Boolean;)V
-	public final fun getAndroidAgpVersion ()Ljava/lang/String;
-	public final fun getAndroidCompileSdk ()Ljava/lang/String;
-	public final fun getAndroidMinSdk ()Ljava/lang/Integer;
-	public final fun getAndroidTargetSdk ()Ljava/lang/Integer;
-	public final fun getApolloGenerateSourcesDuringGradleSync ()Ljava/lang/Boolean;
-	public final fun getApolloLinkSqlite ()Ljava/lang/Boolean;
-	public final fun getGradleVersion ()Ljava/lang/String;
-	public final fun getServiceNames ()Ljava/util/Set;
-}
-
-public final class com/apollographql/apollo/compiler/model/ProjectModel$$serializer : kotlinx/serialization/internal/GeneratedSerializer {
-	public static final field INSTANCE Lcom/apollographql/apollo/compiler/model/ProjectModel$$serializer;
-	public fun childSerializers ()[Lkotlinx/serialization/KSerializer;
-	public fun deserialize (Lkotlinx/serialization/encoding/Decoder;)Lcom/apollographql/apollo/compiler/model/ProjectModel;
-	public synthetic fun deserialize (Lkotlinx/serialization/encoding/Decoder;)Ljava/lang/Object;
-	public fun getDescriptor ()Lkotlinx/serialization/descriptors/SerialDescriptor;
-	public fun serialize (Lkotlinx/serialization/encoding/Encoder;Lcom/apollographql/apollo/compiler/model/ProjectModel;)V
-	public synthetic fun serialize (Lkotlinx/serialization/encoding/Encoder;Ljava/lang/Object;)V
-	public fun typeParametersSerializers ()[Lkotlinx/serialization/KSerializer;
-}
-
-public final class com/apollographql/apollo/compiler/model/ProjectModel$Companion {
-	public final fun serializer ()Lkotlinx/serialization/KSerializer;
-}
-
-public final class com/apollographql/apollo/compiler/model/ProjectModelKt {
-	public static final fun toProjectModel (Ljava/io/File;)Lcom/apollographql/apollo/compiler/model/ProjectModel;
-	public static final fun writeTo (Lcom/apollographql/apollo/compiler/model/ProjectModel;Ljava/io/File;)V
-}
-
-public final class com/apollographql/apollo/compiler/model/ServiceModel {
-	public static final field Companion Lcom/apollographql/apollo/compiler/model/ServiceModel$Companion;
-	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/util/Set;Ljava/util/Set;Ljava/util/Set;Ljava/util/Set;Ljava/lang/String;Ljava/util/Map;Ljava/util/Set;)V
-	public final fun getDownstreamGradleProjectPaths ()Ljava/util/Set;
-	public final fun getEndpointHeaders ()Ljava/util/Map;
-	public final fun getEndpointUrl ()Ljava/lang/String;
-	public final fun getGradleProjectPath ()Ljava/lang/String;
-	public final fun getGraphqlSrcDirs ()Ljava/util/Set;
-	public final fun getSchemaFiles ()Ljava/util/Set;
-	public final fun getServiceName ()Ljava/lang/String;
-	public final fun getTelemetryUsedOptions ()Ljava/util/Set;
-	public final fun getUpstreamGradleProjectPaths ()Ljava/util/Set;
-}
-
-public final class com/apollographql/apollo/compiler/model/ServiceModel$$serializer : kotlinx/serialization/internal/GeneratedSerializer {
-	public static final field INSTANCE Lcom/apollographql/apollo/compiler/model/ServiceModel$$serializer;
-	public fun childSerializers ()[Lkotlinx/serialization/KSerializer;
-	public fun deserialize (Lkotlinx/serialization/encoding/Decoder;)Lcom/apollographql/apollo/compiler/model/ServiceModel;
-	public synthetic fun deserialize (Lkotlinx/serialization/encoding/Decoder;)Ljava/lang/Object;
-	public fun getDescriptor ()Lkotlinx/serialization/descriptors/SerialDescriptor;
-	public fun serialize (Lkotlinx/serialization/encoding/Encoder;Lcom/apollographql/apollo/compiler/model/ServiceModel;)V
-	public synthetic fun serialize (Lkotlinx/serialization/encoding/Encoder;Ljava/lang/Object;)V
-	public fun typeParametersSerializers ()[Lkotlinx/serialization/KSerializer;
-}
-
-public final class com/apollographql/apollo/compiler/model/ServiceModel$Companion {
-	public final fun serializer ()Lkotlinx/serialization/KSerializer;
-}
-
-public final class com/apollographql/apollo/compiler/model/ServiceModelKt {
-	public static final fun toServiceModel (Ljava/io/File;)Lcom/apollographql/apollo/compiler/model/ServiceModel;
-	public static final fun writeTo (Lcom/apollographql/apollo/compiler/model/ServiceModel;Ljava/io/File;)V
-}
-
 public final class com/apollographql/apollo/compiler/operationoutput/OperationDescriptor {
 	public static final field Companion Lcom/apollographql/apollo/compiler/operationoutput/OperationDescriptor$Companion;
 	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
