+++ libraries/apollo-tooling/api/apollo-tooling.api
@@ -126,3 +126,43 @@ public final class com/apollographql/apollo/tooling/Telemetry$TelemetryProperty 
 	public final fun getType ()Ljava/lang/String;
 }
 
+public final class com/apollographql/apollo/tooling/model/ProjectModel {
+	public fun <init> (Ljava/util/Set;Lcom/apollographql/apollo/tooling/model/ProjectModel$TelemetryData;)V
+	public final fun getServiceNames ()Ljava/util/Set;
+	public final fun getTelemetryData ()Lcom/apollographql/apollo/tooling/model/ProjectModel$TelemetryData;
+}
+
+public final class com/apollographql/apollo/tooling/model/ProjectModel$TelemetryData {
+	public fun <init> (Ljava/lang/String;Ljava/lang/Integer;Ljava/lang/Integer;Ljava/lang/String;Ljava/lang/String;Ljava/lang/Boolean;Ljava/lang/Boolean;)V
+	public final fun getAndroidAgpVersion ()Ljava/lang/String;
+	public final fun getAndroidCompileSdk ()Ljava/lang/String;
+	public final fun getAndroidMinSdk ()Ljava/lang/Integer;
+	public final fun getAndroidTargetSdk ()Ljava/lang/Integer;
+	public final fun getApolloGenerateSourcesDuringGradleSync ()Ljava/lang/Boolean;
+	public final fun getApolloLinkSqlite ()Ljava/lang/Boolean;
+	public final fun getGradleVersion ()Ljava/lang/String;
+}
+
+public final class com/apollographql/apollo/tooling/model/ProjectModelKt {
+	public static final fun toProjectModel (Ljava/io/File;)Lcom/apollographql/apollo/tooling/model/ProjectModel;
+	public static final fun writeTo (Lcom/apollographql/apollo/tooling/model/ProjectModel;Ljava/io/File;)V
+}
+
+public final class com/apollographql/apollo/tooling/model/ServiceModel {
+	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/util/Set;Ljava/util/Set;Ljava/util/Set;Ljava/util/Set;Ljava/lang/String;Ljava/util/Map;Ljava/util/Set;)V
+	public final fun getDownstreamGradleProjectPaths ()Ljava/util/Set;
+	public final fun getEndpointHeaders ()Ljava/util/Map;
+	public final fun getEndpointUrl ()Ljava/lang/String;
+	public final fun getGradleProjectPath ()Ljava/lang/String;
+	public final fun getGraphqlSrcDirs ()Ljava/util/Set;
+	public final fun getSchemaFiles ()Ljava/util/Set;
+	public final fun getServiceName ()Ljava/lang/String;
+	public final fun getTelemetryUsedOptions ()Ljava/util/Set;
+	public final fun getUpstreamGradleProjectPaths ()Ljava/util/Set;
+}
+
+public final class com/apollographql/apollo/tooling/model/ServiceModelKt {
+	public static final fun toServiceModel (Ljava/io/File;)Lcom/apollographql/apollo/tooling/model/ServiceModel;
+	public static final fun writeTo (Lcom/apollographql/apollo/tooling/model/ServiceModel;Ljava/io/File;)V
+}
+
