+++ libraries/apollo-tooling/src/main/kotlin/com/apollographql/apollo/tooling/model/ProjectModel.kt
@@ -1,4 +1,4 @@
-package com.apollographql.apollo.compiler.model
+package com.apollographql.apollo.tooling.model
 
 import com.apollographql.apollo.annotations.ApolloInternal
 import kotlinx.serialization.Serializable
@@ -10,16 +10,19 @@ import java.io.File
 @Serializable
 class ProjectModel(
     val serviceNames: Set<String>,
-
-    // Telemetry
-    val gradleVersion: String?,
-    val androidMinSdk: Int?,
-    val androidTargetSdk: Int?,
-    val androidCompileSdk: String?,
-    val androidAgpVersion: String?,
-    val apolloGenerateSourcesDuringGradleSync: Boolean?,
-    val apolloLinkSqlite: Boolean?,
-)
+    val telemetryData: TelemetryData,
+) {
+  @Serializable
+  class TelemetryData(
+      val gradleVersion: String?,
+      val androidMinSdk: Int?,
+      val androidTargetSdk: Int?,
+      val androidCompileSdk: String?,
+      val androidAgpVersion: String?,
+      val apolloGenerateSourcesDuringGradleSync: Boolean?,
+      val apolloLinkSqlite: Boolean?,
+  )
+}
 
 @ApolloInternal
 fun ProjectModel.writeTo(file: File) {
